refactor: handleradapter package 분리
